Update deps to match Angular 18 template
Update dependencies to match what a new Angular 18 workspace would
get as generated.

This includes updating eslint to version 9 and eslint configuration to
its new format. To achieve that there are two dependencies that had to
be listed explicitly in order to get their alpha versions for eslint 9
compatibility. These were also part of the new generated Angular 18
workspace.