Fix peer dependency errors in dependency chains
Yarn reports all peer dependency changes, whether each dependency
fulfills the peer dependency requirements of its own dependencies and
whether they are fulfilled in the dependency tree as a whole.

In this commit, each dependency is "fixed" using packageExtensions
feature of yarn to reflect the missing peer dependency declarations
up in the dependency chain. And those that weren't already fulfilled are
added as direct dependencies.

The version ranges specified in packageExtensions are intentionally
narrow to ensure that they will get rechecked when dependencies are
updated to avoid hiding potential issues in them when dependencies fix
their dependency declarations.