Remove renovate as dependency
Instead, always use latest renovate for validating config file.
The renovate backend using the config file is updated separately and
thus pinning the version of the config validator doesn't serve actual
purpose.

The tool is also included in the full renovate system and the number of
dependencies it brings in is very large.