Rename local dev scripts
Organize scripts in package.json into groups based on prefix:
- `check:` for scripts that perform some check or validation
- `fix:` for scripts that automatically fix things
- `ng:` for scripts that perform things using ng cli tool

Name the scripts based on the tool used for the action instead of trying
to hide that with a concept like "format". When same tool like ng is
used in multiple scripts, then the name contains extra information about
the activity getting executed.

Add also a separate `check` script that runs all the checks.
And a separate `fix` script that runs all fix: group scripts.